requirements.txt
colorama==0.4.6
et_xmlfile==2.0.0
ImageHash==4.3.2
numpy==2.4.0
openpyxl==3.1.5
pillow==12.0.0
PyWavelets==1.9.0
PyYAML==6.0.3
scipy==1.16.3
setuptools==80.9.0
tqdm==4.67.1
wheel==0.45.1
